refactor: move dashboard utilities to frappe.utils
